+++ src/main/java/com/livinglogic/ul4/Utils.java
@@ -1,5 +1,5 @@
 /*
-** Copyright 2009-2022 by LivingLogic AG, Bayreuth/Germany
+** Copyright 2009-2023 by LivingLogic AG, Bayreuth/Germany
 ** All Rights Reserved
 ** See LICENSE for the license
 */
@@ -197,7 +197,14 @@ public class Utils
 	{
 		if (date == null)
 			return null;
-		return Instant.ofEpochMilli(date.getTime()).atZone(ZoneId.systemDefault()).toLocalDateTime();
+		return date.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime();
+	}
+
+	public static LocalDate toLocalDate(Date date)
+	{
+		if (date == null)
+			return null;
+		return date.toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
 	}
 
 	public static BigInteger powerOfTen(int exponent)