+++ CHANGELOG.rst
@@ -1,3 +1,16 @@
+0.181 (2023-03-03)
+------------------
+
+Add ``Utils.toLocalDate()`` method.
+
+
+0.180 (2023-02-10)
+------------------
+
+Add ``DuplicateParameterException``. Throw it when a parameter with the same
+name is added to a ``Signature`` twice.
+
+
 0.179 (2023-01-31)
 ------------------